chore(sdk): Move list and enum to typedefs, typedefs to root
