+++ packages/grafbase-sdk/src/type.ts
@@ -1,5 +1,5 @@
 import { Field } from './field'
-import { ListDefinition } from './field/list'
+import { ListDefinition } from './typedefs/list'
 import { Interface } from './interface'
 import { CacheDefinition, CacheParams, TypeLevelCache } from './typedefs/cache'
 import { ReferenceDefinition } from './typedefs/reference'
